pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.50.0",
    "discord-py>=2.5.2",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gitpython==3.1.43",
    "gunicorn>=23.0.0",
    "openai>=1.77.0",
    "psycopg2-binary>=2.9.10",
    "pynacl>=1.5.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
]
